requirements.txt
Django==3.1.7
graphene-django==2.14.0
django-cors-headers==3.6.0
django-graphql-jwt==4.2.1